third_party/blink/web_tests/fast/events/frame-programmatic-focus-expected.txt

 

This page tests that frames receive focus events when one of their child elements is programmatically focused, and receive blur events when an element not in that frame is programmatically focused.

<input> focused
<input> blurred
main frame blurred
iframe focused
<input> in iframe focused
<input> in iframe blurred
iframe blurred
main frame focused
<input> focused
